Fix bug on token disable, search screen. Reformat and optimise code and layout

+++ app/src/main/java/com/misterixteam/asmodeus/surfgallery/ui/fragment/home/HomeFragmentContract.kt
@@ -32,7 +32,7 @@ interface HomeFragmentContract {
         }
 
         interface ViewModel {
-            fun updatePicture(token: String, onSuccessful: () -> Unit, onError: () -> Unit)
+            fun updatePicture(token: String, onSuccessful: () -> Unit, onError: (Int) -> Unit)
             fun isPictureEmptyState(): MutableState<Boolean>
         }
     }
